// src/contracts/generated/SubmitParams.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ExecutionPlan } from "./ExecutionPlan.js";
import type { TaskId } from "./TaskId.js";

/**
 * `submit` 参数。
 */
export type SubmitParams = { 
/**
 * 由 TS 编译器生成的任务 id。
 */
task_id: TaskId, 
/**
 * 执行计划。
 */
plan: ExecutionPlan, };